feat: lora logic done with trait LoRa
