+++ src/post.rs
@@ -1,18 +1,19 @@
-use crate::{bme280::BME280Sensor, lora::LoRa, BME280Config, Config, LoRaConfig};
-use anyhow::{anyhow, Result};
-use log::{info, error};
-use std::{path::Path, time::Duration};
 use crate::defines::*;
-use std::net::Ipv4Addr;
+use crate::sx1278::LoRa;
+use crate::SX1278;
+use crate::{bme280::BME280Sensor, BME280Config, Config, LoRaConfig};
+use anyhow::{anyhow, Result};
+use log::{error, info};
 use ping::ping;
-
+use std::net::Ipv4Addr;
+use std::{path::Path, time::Duration};
 
 macro_rules! error_log {
     ($func:expr, $type:expr) => {
         match $func {
             Err(e) => {
                 eprintln!("POST: {} not working: {:?}", $type, e);
-                error!("POST: {} not working: {:?}", $type,  e);
+                error!("POST: {} not working: {:?}", $type, e);
                 false
             }
             Ok(_) => true,
@@ -63,17 +64,22 @@ fn post_lora(lora_config: &LoRaConfig) -> Result<()> {
     if !Path::new(&lora_config.spi_config.spidev_path).exists() {
         eprintln!("[ ERR ] SPI POST");
         error!("[ ERR ] SPI POST");
-        return Err(anyhow!("Spidev path {} doesn't exist!", &lora_config.spi_config.spidev_path));
+        return Err(anyhow!(
+            "Spidev path {} doesn't exist!",
+            &lora_config.spi_config.spidev_path
+        ));
     }
 
-    let mut lora = LoRa::from_config(lora_config)?;
+    let mut lora: Box<dyn LoRa> = match lora_config.chip {
+        Chip::SX1278 => Box::new(SX1278::from_config(lora_config)?),
+    };
     let mut mode = 0;
     lora.spi_read_register(LoRaRegister::OP_MODE, &mut mode)?;
     if mode == 0 {
         eprintln!("[ ERR ] SPI POST");
         error!("[ ERR ] SPI POST");
         return Err(anyhow!("Unable to IO via SPI"));
-    } 
+    }
 
     lora.standby_mode()?;
     lora.reset()?;
@@ -83,7 +89,6 @@ fn post_lora(lora_config: &LoRaConfig) -> Result<()> {
         eprintln!("[ ERR ] SPI POST");
         error!("[ ERR ] SPI POST");
         return Err(anyhow!("Unable to IO via SPI"));
-
     } else if mode != 9 {
         eprintln!("[ ERR ] GPIO POST");
         error!("[ ERR ] GPIO POST");
@@ -103,7 +108,10 @@ fn post_bme(bme_config: &BME280Config) -> Result<()> {
     if !Path::new(&bme_config.i2c_bus_path).exists() {
         eprintln!("[ ERR ] BME280 POST");
         error!("[ ERR ] BME280 POST");
-        return Err(anyhow!("i2c bus path {} doesn't exist!", &bme_config.i2c_bus_path));
+        return Err(anyhow!(
+            "i2c bus path {} doesn't exist!",
+            &bme_config.i2c_bus_path
+        ));
     }
 
     let mut bme = match BME280Sensor::new(bme_config.clone()) {
@@ -111,7 +119,7 @@ fn post_bme(bme_config: &BME280Config) -> Result<()> {
             eprintln!("[ ERR ] BME280 POST");
             error!("[ ERR ] BME280 POST");
             return Err(anyhow!("Failed initializing BME280 Sensor: {}", e));
-        },
+        }
         Ok(bme) => bme,
     };
 
@@ -129,17 +137,24 @@ fn post_bme(bme_config: &BME280Config) -> Result<()> {
 fn post_mqtt() -> Result<()> {
     let ip = std::net::IpAddr::V4(Ipv4Addr::new(1, 1, 1, 1));
     let payload = [1; 24];
-    match ping(ip, Some(Duration::from_secs(1)), Some(128), None, None, Some(&payload)) {
+    match ping(
+        ip,
+        Some(Duration::from_secs(1)),
+        Some(128),
+        None,
+        None,
+        Some(&payload),
+    ) {
         Ok(_) => {
             println!("[ OK ] MQTT POST");
             info!("[ OK ] MQTT POST");
             Ok(())
-        },
+        }
         Err(e) => {
             eprintln!("[ ERR ] MQTT POST");
             error!("[ ERR ] MQTT POST");
             Err(anyhow!("Unable to connect to internet: {:?}", e))
-        },
+        }
     }
 }
